Refactore code.
