+++ ui/src/app/widget/lib/well-log-viewer/well-log-component-list.tpl.html
@@ -1,0 +1,69 @@
+<!--
+
+    Copyright © 2016-2018 The Thingsboard Authors
+    Modifications © 2017-2018 Hashmap, Inc
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+-->
+<md-content class="md-padding" layout="column">
+    <md-card>
+        <div style="margin-left: 2%;margin-right: 2%;">
+            <br>
+            <md-input-container class="md-block">
+                <md-select ng-model="trackDetail.width" placeholder="Track Width">
+                    <md-option ng-value="width" ng-repeat="width in trackWidth">{{ width }}</md-option>
+                </md-select>
+            </md-input-container>
+            <br/>
+            <br>
+            <div>
+                <v-accordion id="accordionA" class="vAccordion--default" control="accordionA" onexpand="expandCallback(index, id)" oncollapse="collapseCallback(index, id)">
+
+                    <v-pane id="{{ ::pane.id }}" ng-repeat="component in trackDetail.component" expanded="pane.isExpanded">
+                        <v-pane-header id="{{ ::pane.id }}-header" style="border:outset;">
+                            <lable class="trackLabel">Component {{component.id}}</lable>
+                            <md-button class="md-icon-button md-primary compBtn"
+                                       aria-label="{{ 'action.remove' | translate }}"
+                            ng-click="removeComponent($event,component.id)">
+                                <md-tooltip md-direction="top">
+                                    {{ 'action.remove' | translate }}
+                                </md-tooltip>
+                                <md-icon aria-label="{{ 'action.delete' | translate }}"
+                                         class="material-icons">
+                                    close
+                                </md-icon>
+                            </md-button>
+                        </v-pane-header>
+
+                        <v-pane-content id="{{ ::pane.id }}-content" aria-labelledby="{{ ::pane.id }}-header">
+                            <tb-well-log-component
+                                    track-component="component"
+                                    datasources="datasources"
+                            ></tb-well-log-component>
+                        </v-pane-content>
+                    </v-pane>
+
+                </v-accordion>
+                <br/>
+                <md-button ng-disabled="$root.loading" ng-click="addComponent($event)" class="md-raised md-primary md-button md-ink-ripple md-themeDark-theme" aria-label="Add Track">
+                    <md-tooltip>
+                        Add Component
+                    </md-tooltip>
+                    Add Component
+
+                </md-button>
+            </div>
+        </div>
+    </md-card>
+</md-content>
