+++ ui/src/app/widget/lib/well-log-viewer/well-log-component.directive.js
@@ -16,42 +16,27 @@
  */
 /* eslint-disable import/no-unresolved, import/default */
 
-import wellTrackComponent from './well-log-component.tpl.html'
+import wellTrackComponents from './well-log-component.tpl.html'
 
 /* eslint-enable import/no-unresolved, import/default */
 
 /*@ngInject*/
-
-
-export default function WellLogViewerTrackDirective($compile, $templateCache) {
+export default function WellLogViewerComponentsDirective($compile, $templateCache) {
     var linker = function (scope, element) {
-        var template = $templateCache.get(wellTrackComponent);
+        var template = $templateCache.get(wellTrackComponents);
         element.html(template);
-        scope.trackWidth =[1,2,3,4,5];
+        scope.componentTypes =["Line", "Linear Grid", "Time Y axis", "Mud Log Viewer"];
+        scope.fillTypes =["left", "right"];
+        scope.styleTypes =["dashed", "solid"];
 
-        scope.addComponent = function (){
-
-            var insertDetail = {
-                id: scope.trackDetail.component.length ? scope.trackDetail.component.length + 1 : 1
-            }
-            scope.trackDetail.component.push(insertDetail);
-        }
-        scope.removeComponent = function ($event,id){
-            var index = scope.trackDetail.component.findIndex(x => x.id==id);
-            if($event){
-                  $event.stopPropagation();
-                  $event.preventDefault();
-            }
-            scope.trackDetail.component.splice(index, 1);
-        }
         $compile(element.contents())(scope);
     }
     return {
         restrict: "E",
         link: linker,
         scope: {
-            trackDetail: '=',
-            datasources: '='
+            trackComponent: '=',
+            datasources:'='
         }
     };
 }
