--- ui/src/app/widget/lib/well-log-viewer/well-log-components.tpl.html
@@ -1,146 +1,0 @@
-<!--
-
-    Copyright © 2016-2018 The Thingsboard Authors
-    Modifications © 2017-2018 Hashmap, Inc
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<md-content class="md-padding" layout="column">
-    <md-card>
-        <div style="margin-left: 2%;margin-right: 2%;">
-<div>
-    <md-input-container class="md-block">
-        <md-select ng-model="trackComponent.cType" placeholder="Select Component">
-            <md-option ng-value="component" ng-repeat="component in componentTypes">{{ component }}</md-option>
-        </md-select>
-    </md-input-container>
-</div>
-<div>
-    <md-input-container class="md-block">
-        <div ng-repeat="datasource in datasources">
-            <!--<div ng-repeat="datakey in datasource.value.dataKeys">-->
-
-                <md-select ng-model="trackComponent.datasource" placeholder="Datasource">
-                    <md-option ng-value="d.label" ng-repeat="d in datasource.value.dataKeys">{{ d.label }}</md-option>
-                </md-select>
-            <!--</div>-->
-        </div>
-    </md-input-container>
-</div>
-<div flex
-     md-color-picker
-     ng-model="trackComponent.color"
-     label="Color"
-     icon="format_color_fill"
-     default="rgba(0, 0, 0, 0.87)"
-     md-color-clear-button="false"
-     open-on-input="true"
-     md-color-generic-palette="false"
-     md-color-history="false"
-></div>
-<div>
-    <md-input-container class="md-block">
-        <label translate>Line Width</label>
-        <input  name="lineWidth" ng-model="trackComponent.lineWidth">
-    </md-input-container>
-</div>
-<div>
-    <md-checkbox ng-model="trackComponent.hasHeader" aria-label="Checkbox 1">
-        Has Header
-    </md-checkbox>
-    <br/>
-    <fieldset ng-if="trackComponent.hasHeader">
-        <md-input-container class="md-block">
-            <label translate>Header Minimum</label>
-            <input  name="name" ng-model="trackComponent.headerMin">
-        </md-input-container>
-        <md-input-container class="md-block">
-            <label translate>Header Maximum</label>
-            <input  name="name" ng-model="trackComponent.headerMax">
-        </md-input-container>
-        <md-input-container class="md-block">
-            <label translate>Header Name</label>
-            <input  name="name" ng-model="trackComponent.headerName">
-        </md-input-container>
-    </fieldset>
-</div>
-<div>
-    <md-checkbox ng-model="trackComponent.areaFill.enable" aria-label="Checkbox 1">
-        Enable Area Fill
-    </md-checkbox>
-    <br/>
-    <fieldset ng-if="trackComponent.areaFill.enable">
-        <md-input-container class="md-block">
-        <md-select ng-model="trackComponent.areaFill.fill" placeholder="Select Area Fill">
-            <md-option ng-value="fill" ng-repeat="fill in fillTypes">{{ fill }}</md-option>
-        </md-select>
-        </md-input-container>
-        <div flex
-             md-color-picker
-             ng-model="trackComponent.areaFill.color"
-             label="Fill Color"
-             icon="format_color_fill"
-             default="rgba(0, 0, 0, 0.87)"
-             md-color-clear-button="false"
-             open-on-input="true"
-             md-color-generic-palette="false"
-             md-color-history="false"
-        ></div>
-        <md-input-container class="md-block">
-            <label translate>Fill opacity</label>
-            <input  name="name" ng-model="trackComponent.areaFill.opacity">
-        </md-input-container>
-    </fieldset>
-</div>
-<div>
-    <md-checkbox ng-model="trackComponent.componentDetail.girdMinorLines" aria-label="Checkbox 1">
-        Grid minor lines
-    </md-checkbox>
-    <br/>
-    <fieldset ng-if="trackComponent.componentDetail.girdMinorLines">
-        <md-input-container class="md-block">
-            <label translate>Number of lines</label>
-            <input  name="name" ng-model="trackComponent.minorLines.lines">
-        </md-input-container>
-        <md-input-container class="md-block">
-            <label translate>Line weight</label>
-            <input  name="name" ng-model="trackComponent.minorLines.lineWeight">
-        </md-input-container>
-        <md-input-container class="md-block">
-            <md-select ng-model="trackComponent.minorLines.style" placeholder="Style">
-                <md-option ng-value="style" ng-repeat="style in styleTypes">{{ style }}</md-option>
-            </md-select>
-        </md-input-container>
-    </fieldset>
-</div>
-<div>
-    <md-checkbox ng-model="trackComponent.componentDetail.girdMajorLines" aria-label="Checkbox 1">
-        Grid major lines
-    </md-checkbox>
-    <br/>
-    <fieldset ng-if="trackComponent.componentDetail.girdMajorLines">
-        <md-input-container class="md-block">
-            <label translate>Number of lines</label>
-            <input  name="name" ng-model="trackComponent.majorLines.lines">
-        </md-input-container>
-        <md-input-container class="md-block">
-            <label translate>Line weight</label>
-            <input  name="name" ng-model="trackComponent.majorLines.lineWeight">
-        </md-input-container>
-    </fieldset>
-</div>
-        </div>
-    </md-card>
-</md-content>
-
