+++ ui/src/app/widget/lib/well-log-viewer/well-log-track.tpl.html
@@ -35,10 +35,10 @@
         </v-pane-header>
 
         <v-pane-content id="{{ ::pane.id }}-content" aria-labelledby="{{ ::pane.id }}-header">
-            <tb-well-log-component
+            <tb-well-log-component-list
                     track-detail="track"
                     datasources = "datasources"
-            ></tb-well-log-component>
+            ></tb-well-log-component-list>
         </v-pane-content>
     </v-pane>
